Core/src/main/java/repositories/MessageRepository.java
package repositories;

import org.hibernate.Session;
import org.hibernate.SessionFactory;
import org.hibernate.Transaction;
import org.hibernate.cfg.Configuration;

public class MessageRepository <T>{

    private static SessionFactory sessionFactory;


    public boolean saveMessage(T messageEntity){
        Transaction transaction = null;
        sessionFactory = new Configuration().configure().buildSessionFactory();
        try (Session session = sessionFactory.openSession()) {
            // start a transaction
            transaction = session.beginTransaction();
            // save the student objects
            session.save(messageEntity);
            // commit transaction
            transaction.commit();

        } catch (Exception e) {
            if (transaction != null) {
                transaction.rollback();
            }
            e.printStackTrace();
            return false;
        }

        return true;
    }


}
